Ensure the storage directory is calculated during runtime
When building nimlsp in a sandbox (before installing it on a system) the
temporary directory might be different. In my case I was building nimlsp
with Nix which caused nimlsp to hardcode `/build/nimlsp` as storage
directory. That directory only exists within the build sandbox and even
if it would exist outside of the sandbox it wouldn't be writable by the
average user.

Side note: This is the very first Nim code I've ever "written".