fix(nimlsp): move Nim version check to run after initialize (#182)
* fix(nimlsp): move Nim version check to run after initialize

Previously,  was called at the start of , before the
client had sent an initialize request. This caused potential protocol
violations or confused clients.

Now the version check is deferred to run only after receiving and handling
the initialize request, ensuring it adheres to the LSP handshake and
avoids premature messages.

* upgrade actions/cache version from 1 to 4

* remvoe nim 1.2 and 1.4 test